--- a/qcodes/tests/dataset/test_dataset_basic.py
+++ b/qcodes/tests/dataset/test_dataset_basic.py
@@ -24,7 +24,11 @@ from qcodes.dataset.descriptions.param_spec import ParamSpecBase
 from qcodes.dataset.descriptions.rundescriber import RunDescriber
 from qcodes.dataset.guids import parse_guid
 from qcodes.dataset.sqlite.connection import atomic, path_to_dbfile
-from qcodes.dataset.sqlite.database import _convert_array, get_DB_location
+from qcodes.dataset.sqlite.database import (
+    _convert_array,
+    _convert_complex,
+    get_DB_location,
+)
 from qcodes.dataset.sqlite.queries import _rewrite_timestamps, _unicode_categories
 from qcodes.tests.common import error_caused_by
 from qcodes.tests.dataset.helper_functions import verify_data_dict
@@ -613,6 +617,15 @@ def test_backward_compat__adapt_array_v0_33():
         np.save(out, arr)
         out.seek(0)
         assert arr == _convert_array(out.read())
+
+
+def test_backward_compat__adapt_complex_v0_33():
+    for dtype in complex_types:
+        value = dtype(1j)
+        out = io.BytesIO()
+        np.save(out, np.array([value]))
+        out.seek(0)
+        assert value == _convert_complex(out.read())
 
 
 def test_missing_keys(dataset):